# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Benim Lezzet Dünyam — Sipariş API
  version: 1.0.0
  summary: Catering sipariş platformunun normatif API sözleşmesi.
  description: |
    Bu dosya **normatif sözleşmedir**. `docs/03-api-sozlesmesi.md` aynı sözleşmenin
    insan için yazılmış açıklamasıdır; ikisi çelişirse **bu dosya kazanır**.

    `packages/api_client` ve `website/lib/api` bu dosyadan üretilir.
    `platform/` aynı sözleşmeyi `openapi.json` olarak üretir; CI ikisi arasındaki
    anlamsal farkı kontrol eder ve fark varsa build kırmızıdır.

    ## Değişmezler
    - **Para** kuruş cinsinden tam sayıdır. `4550` = 45,50 TL. Ondalık float asla kullanılmaz.
    - **Tarih/saat** ISO 8601 UTC'dir. İstemci Europe/Istanbul'a çevirir.
    - **Uyumluluk:** alan eklenebilir; silinemez, adı/tipi değişemez. İstemciler bilmedikleri
      alanları yok sayar. Kırıcı değişiklik `/api/v2/` ile açılır.
    - **Kapsam dışı:** öğrenci ve kurum içi sipariş kanalları yoktur ve olmayacaktır
      (`docs/00-genel-bakis.md` §4). Sipariş çeşitliliği yalnızca `delivery_type` ile ifade edilir.

    ## Oran sınırları
    | Uç grubu | Sınır |
    |---|---|
    | `/auth/*` | 10 istek / dakika / IP |
    | `POST /orders` | 20 istek / saat / hesap |
    | `/kitchen/*` | 1200 istek / saat / cihaz (5 sn polling = 720/saat, üstüne heartbeat ve durum trafiği) |
    | Diğer | 120 istek / dakika / IP |

    Sınır aşılırsa `429 RATE_LIMITED`.
  license:
    name: Tescilli — dahili kullanım
...
servers:
...
    description: Prod
...
security:
  - CustomerToken: []
paths:
...
  /auth/me:
    parameters: [$ref: '#/components/parameters/AppId', $ref: '#/components/parameters/AppVersion', $ref: '#/components/parameters/AcceptLanguage']
    get:
      tags: [Kimlik]
      summary: Oturumdaki müşteri
      operationId: me
      responses:
        '200':
          description: Müşteri profili
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  # ─────────────────────────────── Katalog ──────────────────────────────
    patch:
      tags: [Kimlik]
      summary: Profil güncelle
      description: |
        Ad, soyad ve telefon güncellenir. Gönderilmeyen alanlar değişmez.

        **E-posta DEĞİŞTİRİLEMEZ.** Giriş kimliği odur; değiştirmek yeni
        adrese onay bağlantısı göndermeyi gerektirir, o olmadan hesabı
        başkasının e-postasına taşımanın yolu olur. Gövdede `email`
        gönderilirse sessizce yok sayılır.
      operationId: updateMe
      security:
        - CustomerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string, maxLength: 64 }
                last_name: { type: string, maxLength: 64 }
                telephone:
                  type: string
                  pattern: '^[1-9][0-9]{9}$'
                  description: 10 hane, başında 0 veya +90 olmadan.
      responses:
        '200':
          description: Güncel profil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422':
          description: Doğrulama hatası
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/password:
    parameters: [$ref: '#/components/parameters/AppId', $ref: '#/components/parameters/AppVersion', $ref: '#/components/parameters/AcceptLanguage']
    post:
      tags: [Kimlik]
      summary: Parola değiştir
      description: |
        **Mevcut parola zorunludur.** Token'ı çalınmış bir oturumun
        parolayı değiştirip hesabı tamamen ele geçirmesini engelleyen tek
        şey budur.

        Başarılı olduğunda **diğer tüm token'lar iptal edilir** ve çağırana
        yeni bir token verilir. Parola değiştirmenin amacı "başkası
        giremesin"dir; eski oturumları açık bırakmak o amacı boşa çıkarır.

        `bld-auth` oran sınırına tabidir (dakikada 10): mevcut parolayı
        denemek kaba kuvvete açıktır.
      operationId: changePassword
      security:
        - CustomerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password]
              properties:
                current_password: { type: string }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200':
          description: Değiştirildi; yeni token döner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422':
          description: Mevcut parola yanlış veya doğrulama hatası
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    CustomerToken:
      type: http
      scheme: bearer
      description: Müşteri hesabı token'ı (`customer` kapsamı). `/kitchen/*` uçlarına erişemez.
    KitchenToken:
      type: http
      scheme: bearer
      description: |
        Cihaz token'ı (`kitchen` kapsamı). `POST /kitchen/pair` ile alınır.
        Fiyat, müşteri iletişim bilgisi ve rapor uçlarına erişemez.
  parameters:
    AppId:
      name: X-App-Id
      in: header
      required: true
      schema: { type: string, enum: [website, musteriapp, mutfakapp] }
    AppVersion:
      name: X-App-Version
      in: header
      required: true
      schema: { $ref: '#/components/schemas/SemVer' }
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: true
      schema: { type: string, default: tr }
...
  responses:
    Unauthenticated:
      description: Token yok veya geçersiz (`UNAUTHENTICATED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Çok fazla istek (`RATE_LIMITED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    SemVer:
      type: string
      pattern: '^\d+\.\d+\.\d+$'
      examples: ['1.0.0']
...
    ErrorCode:
      type: string
      enum:
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_FAILED
        - INVALID_TRANSITION
        - LOCATION_CLOSED
        - ITEM_UNAVAILABLE
        - DEVICE_REVOKED
        - RATE_LIMITED
        - SERVER_ERROR
    Error:
      type: object
      required: [error]
      description: Tüm hataların tek biçimi.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message:
              type: string
              description: Kullanıcıya gösterilebilir Türkçe metin.
            details:
              type: [object, 'null']
              additionalProperties: true
      examples:
        - error:
            code: INVALID_TRANSITION
            message: Sipariş bu duruma geçirilemez.
            details: { from: yeni, to: hazir }
...
    AuthResponse:
      type: object
      required: [token, customer]
      properties:
        token: { type: string }
        customer:
          type: object
          required: [id, first_name]
          properties:
            id: { type: integer, format: int64 }
            first_name: { type: string }
    Customer:
      type: object
      required: [id, first_name, last_name, email, telephone]
      properties:
        id: { type: integer, format: int64 }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        telephone: { type: string }
        default_location_id: { type: [integer, 'null'], format: int64 }
        account_type:
          type: string
          enum: [corporate, individual]
          description: B2B — yeni kayıtlar `corporate`. Bireysel hesap sipariş veremez.
        can_order:
          type: boolean
          description: İstemci sipariş yolunu bununla açar/kapatır; sunucu da denetler.
        company_name: { type: [string, 'null'] }
        contact_person: { type: [string, 'null'] }

    # ── Cari hesap ──────────────────────────────────────────────────────
...