# platform/extensions/veykemtu/bridgeapi/src/Http/Controllers/AuthController.php
<?php

declare(strict_types=1);

namespace Veykemtu\BridgeApi\Http\Controllers;

use Igniter\User\Models\CustomerGroup;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Hash;
use Illuminate\Validation\Rule;
use Veykemtu\BridgeApi\Exceptions\ApiException;
use Veykemtu\BridgeApi\Models\ApiCustomer;
use Veykemtu\BridgeApi\Services\CustomerGate;

/**
 * Kimlik uçları — `docs/openapi.yaml` §Kimlik.
 *
 * Token'lar Sanctum'dur ve `customer` yeteneğiyle üretilir; mutfak uçlarına
 * geçemezler (ADR-08, `docs/10-test-kabul.md` S5).
 */
class AuthController extends ApiController
{
    public function register(Request $request): JsonResponse
    {
        $data = $request->validate([
            'first_name' => ['required', 'string', 'max:64'],
            'last_name' => ['required', 'string', 'max:64'],
            'email' => ['required', 'email', 'max:96', Rule::unique('customers', 'email')],
            'telephone' => ['required', 'string', 'regex:/^[1-9][0-9]{9}$/'],
            'password' => ['required', 'string', 'min:8', 'max:128'],
            'kvkk_accepted' => ['required', 'accepted'],

            // Kurumsal (B2B) alanları — ŞİMDİLİK OPSİYONEL. Web ve mobil kayıt
            // formları bu alanları göndermeye başlayana kadar zorunlu kılmak
            // mevcut web kaydını 422 ile kırardı (ADR-09 kademeli geçiş).
            // Sunucu her yeni kaydı yine de 'corporate' işaretler; alanlar
            // dolunca (Aşama B) zorunluya çevrilecek.
            'company_name' => ['sometimes', 'nullable', 'string', 'max:160'],
            'tax_office' => ['sometimes', 'nullable', 'string', 'max:120'],
            'tax_number' => ['sometimes', 'nullable', 'string', 'max:32'],
            'contact_person' => ['sometimes', 'nullable', 'string', 'max:120'],
            'company_phone' => ['sometimes', 'nullable', 'string', 'max:32'],
        ], [
            'kvkk_accepted.accepted' => 'KVKK aydınlatma metnini onaylamanız gerekiyor.',
            'telephone.regex' => 'Telefon 10 haneli olmalı (başında 0 veya +90 olmadan).',
            'email.unique' => 'Bu e-posta zaten kayıtlı.',
        ]);

        $customer = new ApiCustomer;
        $customer->first_name = $data['first_name'];
        $customer->last_name = $data['last_name'];
        $customer->email = $data['email'];
        $customer->telephone = $data['telephone'];
        $customer->password = $data['password'];
        $customer->customer_group_id = $this->defaultGroupId();
        // B2B: her yeni kayıt kurumsaldır; bireysel self-servis kayıt yok.
        // Alanlar tek tek atanır (toplu fill değil) — çekirdek Customer'ın
        // fillable'ı bld_* kolonlarını içermez, AuthController kalıbı budur.
        $customer->bld_account_type = CustomerGate::TYPE_CORPORATE;
        $customer->bld_org_name = $data['company_name'] ?? null;
        $customer->bld_tax_office = $data['tax_office'] ?? null;
        $customer->bld_tax_no = $data['tax_number'] ?? null;
        $customer->bld_contact_person = $data['contact_person'] ?? null;
        $customer->bld_org_phone = $data['company_phone'] ?? null;
        $customer->status = true;
        // KVKK onayı zaman damgasıyla saklanır (docs/02 §6).
        $customer->is_activated = true;
        $customer->activated_at = now();
        $customer->save();

        return $this->json($this->authPayload($customer), 201);
    }

    public function login(Request $request): JsonResponse
    {
        $data = $request->validate([
            'email' => ['required', 'email'],
            'password' => ['required', 'string'],
        ]);

        $customer = ApiCustomer::where('email', $data['email'])->first();

        // Kullanıcı yok ile şifre yanlış aynı mesajı döner: hangi e-postaların
        // kayıtlı olduğunu sızdırmamak için.
        if ($customer === null || !Hash::check($data['password'], (string) $customer->password)) {
            throw ApiException::validationFailed('E-posta veya şifre hatalı.', [
                'email' => 'Doğrulanamadı.',
            ]);
        }

        if ((bool) $customer->status !== true) {
            throw ApiException::forbidden('Hesabınız devre dışı. Bizimle iletişime geçin.');
        }

        return $this->json($this->authPayload($customer));
    }

    public function logout(Request $request): JsonResponse
    {
        // Yalnızca bu isteği yapan token silinir; kullanıcının diğer
        // cihazlardaki oturumları etkilenmez.
        $request->user()?->currentAccessToken()?->delete();

        return $this->noContent();
    }

    public function me(Request $request): JsonResponse
    {
        /** @var ApiCustomer $customer */
        $customer = $request->user();

        $accountType = (string) ($customer->bld_account_type ?? CustomerGate::TYPE_CORPORATE);

        return $this->json([
            'id' => (int) $customer->customer_id,
            'first_name' => (string) $customer->first_name,
            'last_name' => (string) $customer->last_name,
            'email' => (string) $customer->email,
            'telephone' => (string) ($customer->telephone ?? ''),
            'default_location_id' => $this->defaultLocationId(),
            // Kurumsal (B2B) alanları — additive. can_order istemci gating'i
            // için: bireysel işaretli hesap sipariş veremez (sunucu da engeller).
            'account_type' => $accountType,
            'can_order' => $accountType === CustomerGate::TYPE_CORPORATE,
            'company_name' => $customer->bld_org_name,
            'contact_person' => $customer->bld_contact_person,
        ]);
    }

    /**
     * Profil güncelleme.
     *
     * E-POSTA DEĞİŞTİRİLEMEZ. Giriş kimliği odur; değiştirmek doğrulama
     * akışı (yeni adrese onay bağlantısı) ister, o da yoksa hesabı
     * başkasının e-postasına taşımanın yolu olur. Ayrı bir iş.
     */
    public function updateMe(Request $request): JsonResponse
    {
        /** @var ApiCustomer $customer */
        $customer = $request->user();

        $data = $request->validate([
            'first_name' => ['sometimes', 'required', 'string', 'max:64'],
            'last_name' => ['sometimes', 'required', 'string', 'max:64'],
            'telephone' => ['sometimes', 'required', 'string', 'regex:/^[1-9][0-9]{9}$/'],
            // Kurumsal alanlar güncellenebilir; vergi kimliği (tax_no/office)
            // e-posta gibi kilitli — değişimi ayrı doğrulama ister.
            'company_name' => ['sometimes', 'nullable', 'string', 'max:160'],
            'contact_person' => ['sometimes', 'nullable', 'string', 'max:120'],
            'company_phone' => ['sometimes', 'nullable', 'string', 'max:32'],
        ], [
            'telephone.regex' => 'Telefon 10 haneli olmalı (başında 0 veya +90 olmadan).',
        ]);

        foreach (['first_name', 'last_name', 'telephone'] as $alan) {
            if (array_key_exists($alan, $data)) {
                $customer->{$alan} = $data[$alan];
            }
        }

        // Kurumsal alanlar `bld_` önekli kolonlara eşlenir.
        $kurumsalEsleme = [
            'company_name' => 'bld_org_name',
            'contact_person' => 'bld_contact_person',
            'company_phone' => 'bld_org_phone',
        ];
        foreach ($kurumsalEsleme as $gelen => $kolon) {
            if (array_key_exists($gelen, $data)) {
                $customer->{$kolon} = $data[$gelen];
            }
        }

        $customer->save();

        return $this->me($request);
    }

    /**
     * Parola değiştirme.
     *
     * MEVCUT PAROLA ZORUNLU. Token çalınmış bir oturumun parolayı
     * değiştirip hesabı tamamen ele geçirmesini engelleyen tek şey bu.
     *
     * Değişiklikten sonra DİĞER TÜM TOKEN'LAR İPTAL EDİLİR ve çağırana
     * yenisi verilir: parola değiştirmenin amacı zaten "başkası
     * giremesin"dir; eski oturumları açık bırakmak o amacı boşa çıkarır.
     */
    public function changePassword(Request $request): JsonResponse
    {
        /** @var ApiCustomer $customer */
        $customer = $request->user();

        $data = $request->validate([
            'current_password' => ['required', 'string'],
            'password' => ['required', 'string', 'min:8', 'max:128', 'different:current_password'],
        ], [
            'password.different' => 'Yeni parola eskisiyle aynı olamaz.',
        ]);

        if (!Hash::check($data['current_password'], (string) $customer->password)) {
            throw ApiException::validationFailed('Mevcut parola yanlış.', [
                'current_password' => 'Mevcut parola yanlış.',
            ]);
        }

        $customer->password = $data['password'];
        $customer->save();

        $customer->tokens()->delete();

        return $this->json($this->authPayload($customer));
    }

    public function pushToken(Request $request): JsonResponse
    {
        $data = $request->validate([
            'fcm_token' => ['required', 'string', 'max:512'],
            'platform' => ['sometimes', Rule::in(['android'])],
        ]);

        /** @var ApiCustomer $customer */
        $customer = $request->user();

        // Tablo `veykemtu/push` eklentisine aittir ama uç sözleşmede müşteri
        // tarafındadır. Eklenti henüz yoksa istek sessizce başarılı sayılır:
        // istemcinin push kaydı yüzünden hata alması, siparişten daha
        // önemsiz bir işin ana akışı bozması demek olurdu.
        if (!$this->pushTableExists()) {
            return $this->noContent();
        }

        DB::table('veykemtu_device_tokens')->updateOrInsert(
            ['customer_id' => $customer->customer_id, 'fcm_token' => $data['fcm_token']],
            ['platform' => $data['platform'] ?? 'android', 'updated_at' => now()],
        );

        return $this->noContent();
    }

    /** @return array<string, mixed> */
    private function authPayload(ApiCustomer $customer): array
    {
        // Aynı hesabın birden çok cihazda açık kalması normaldir (web + mobil),
        // bu yüzden eski token'lar silinmez.
        $token = $customer->createToken('istemci', [ApiCustomer::SCOPE]);

        return [
            'token' => $token->plainTextToken,
            'customer' => [
                'id' => (int) $customer->customer_id,
                'first_name' => (string) $customer->first_name,
            ],
        ];
    }

    private function defaultGroupId(): ?int
    {
        $id = CustomerGroup::where('group_name', 'Catering Müşterisi')->value('customer_group_id');

        return $id !== null ? (int) $id : null;
    }

    private function defaultLocationId(): ?int
    {
        $id = DB::table('locations')
            ->where('location_status', true)
            ->orderByDesc('is_default')
            ->value('location_id');

        return $id !== null ? (int) $id : null;
    }

    private function pushTableExists(): bool
    {
        return DB::getSchemaBuilder()->hasTable('veykemtu_device_tokens');
    }
}